+++ pkgs/native_assets_builder/test/build_runner/build_runner_build_dry_run_test.dart
@@ -28,6 +28,7 @@ void main() async {
         packageUri,
         logger,
         dartExecutable,
+        hasLinkPhase: false,
       );
       final dryRunAssets = dryRunResult.assets.toList();
       final result = await build(