+++ pkgs/native_assets_builder/test/build_runner/build_runner_cycle_test.dart
@@ -26,6 +26,7 @@ void main() async {
           packageUri,
           createCapturingLogger(logMessages, level: Level.SEVERE),
           dartExecutable,
+          hasLinkPhase: false,
         );
         final fullLog = logMessages.join('\n');
         expect(result.success, false);