+++ pkgs/native_assets_builder/test/build_runner/build_runner_reusability_test.dart
@@ -33,12 +33,14 @@ void main() async {
         linkModePreference: LinkModePreferenceImpl.dynamic,
         workingDirectory: packageUri,
         includeParentEnvironment: true,
+        hasLinkPhase: false,
       );
       await buildRunner.buildDryRun(
         targetOS: Target.current.os,
         linkModePreference: LinkModePreferenceImpl.dynamic,
         workingDirectory: packageUri,
         includeParentEnvironment: true,
+        hasLinkPhase: false,
       );
       await buildRunner.build(
         buildMode: BuildModeImpl.release,
@@ -46,6 +48,7 @@ void main() async {
         target: Target.current,
         workingDirectory: packageUri,
         includeParentEnvironment: true,
+        hasLinkPhase: false,
       );
       await buildRunner.build(
         buildMode: BuildModeImpl.release,
@@ -53,6 +56,7 @@ void main() async {
         target: Target.current,
         workingDirectory: packageUri,
         includeParentEnvironment: true,
+        hasLinkPhase: false,
       );
     });
   });