+++ pkgs/native_assets_cli/lib/src/model/build_config.dart
@@ -34,6 +34,19 @@ final class BuildConfigImpl extends HookConfigImpl implements BuildConfig {
 
   final Map<String, Metadata>? _dependencyMetadata;
 
+  @override
+  bool get hasLinkPhase {
+    if (version <= Version(1, 2, 0)) {
+      return false;
+    }
+    if (version == Version(1, 3, 0)) {
+      return true;
+    }
+    return _hasLinkPhase as bool;
+  }
+
+  final bool? _hasLinkPhase;
+
   static List<String> _supportedAssetTypesBackwardsCompatibility(
     Iterable<String>? supportedAssetTypes,
   ) =>
@@ -55,14 +68,22 @@ final class BuildConfigImpl extends HookConfigImpl implements BuildConfig {
     required super.targetOS,
     required super.linkModePreference,
     Map<String, Metadata>? dependencyMetadata,
+    required bool? hasLinkPhase,
     super.dryRun,
   })  : _dependencyMetadata = dependencyMetadata,
+        _hasLinkPhase = hasLinkPhase,
         super(
           hook: Hook.build,
           version: version ?? HookConfigImpl.latestVersion,
           supportedAssetTypes:
               _supportedAssetTypesBackwardsCompatibility(supportedAssetTypes),
-        );
+        ) {
+    if (this.version < Version(1, 4, 0)) {
+      assert(hasLinkPhase == null);
+    } else {
+      assert(hasLinkPhase != null);
+    }
+  }
 
   BuildConfigImpl.dryRun({
     required super.outputDirectory,
@@ -70,8 +91,10 @@ final class BuildConfigImpl extends HookConfigImpl implements BuildConfig {
     required super.packageRoot,
     required super.targetOS,
     required super.linkModePreference,
+    required bool? hasLinkPhase,
     Iterable<String>? supportedAssetTypes,
   })  : _dependencyMetadata = null,
+        _hasLinkPhase = hasLinkPhase,
         super.dryRun(
           hook: Hook.build,
           version: HookConfigImpl.latestVersion,
@@ -100,6 +123,8 @@ final class BuildConfigImpl extends HookConfigImpl implements BuildConfig {
 
   static const dependencyMetadataConfigKey = 'dependency_metadata';
 
+  static const hasLinkPhaseKey = 'has_link_phase';
+
   static BuildConfigImpl _readFieldsFromConfig(Config config) {
     final dryRun = HookConfigImpl.parseDryRun(config) ?? false;
     final targetOS = HookConfigImpl.parseTargetOS(config);
@@ -113,6 +138,7 @@ final class BuildConfigImpl extends HookConfigImpl implements BuildConfig {
           HookConfigImpl.parseTargetArchitecture(config, dryRun, targetOS),
       linkModePreference: HookConfigImpl.parseLinkModePreference(config),
       dependencyMetadata: parseDependencyMetadata(config),
+      hasLinkPhase: parseHasLinkPhase(config),
       version: HookConfigImpl.parseVersion(config),
       cCompiler: HookConfigImpl.parseCCompiler(config, dryRun),
       supportedAssetTypes: HookConfigImpl.parseSupportedAssetTypes(config),
@@ -127,11 +153,7 @@ final class BuildConfigImpl extends HookConfigImpl implements BuildConfig {
     );
   }
 
-  static Map<String, Metadata>? parseDependencyMetadata(Config config) =>
-      _readDependencyMetadataFromConfig(config);
-
-  static Map<String, Metadata>? _readDependencyMetadataFromConfig(
-      Config config) {
+  static Map<String, Metadata>? parseDependencyMetadata(Config config) {
     final fileValue =
         config.valueOf<Map<Object?, Object?>?>(dependencyMetadataConfigKey);
     if (fileValue == null) {
@@ -156,6 +178,9 @@ final class BuildConfigImpl extends HookConfigImpl implements BuildConfig {
     ).sortOnKey();
   }
 
+  static bool? parseHasLinkPhase(Config config) =>
+      config.optionalBool(hasLinkPhaseKey);
+
   static BuildConfigImpl fromJson(Map<String, dynamic> buildConfigJson) =>
       BuildConfigImpl._fromConfig(Config(fileParsed: buildConfigJson));
 
@@ -169,6 +194,7 @@ final class BuildConfigImpl extends HookConfigImpl implements BuildConfig {
                   MapEntry(packageName, metadata.toJson()),
             ),
         },
+        if (version >= Version(1, 4, 0)) hasLinkPhaseKey: hasLinkPhase,
       }.sortOnKey();
 
   @override
@@ -184,6 +210,9 @@ final class BuildConfigImpl extends HookConfigImpl implements BuildConfig {
             .equals(other._dependencyMetadata, _dependencyMetadata)) {
       return false;
     }
+    if (_hasLinkPhase != other._hasLinkPhase) {
+      return false;
+    }
     return true;
   }
 
@@ -191,6 +220,7 @@ final class BuildConfigImpl extends HookConfigImpl implements BuildConfig {
   int get hashCode => Object.hashAll([
         super.hashCode,
         linkModePreference,
+        hasLinkPhase,
         if (!dryRun) ...[
           const DeepCollectionEquality().hash(_dependencyMetadata),
         ],