+++ pkgs/native_assets_cli/test/model/build_output_test.dart
@@ -213,6 +213,7 @@ version: 1.0.0''';
       targetArchitecture: ArchitectureImpl.arm64,
       targetOS: OSImpl.macOS,
       linkModePreference: LinkModePreferenceImpl.dynamic,
+      hasLinkPhase: false,
     );
     final buildOutput = getBuildOutput();
     await buildOutput.writeToFile(config: config);
@@ -234,6 +235,7 @@ version: 1.0.0''';
       targetOS: OSImpl.macOS,
       linkModePreference: LinkModePreferenceImpl.dynamic,
       version: Version(1, 1, 0),
+      hasLinkPhase: null, // version < 1.4.0
     );
     final buildOutput = getBuildOutput(withLinkedAssets: false);
     await buildOutput.writeToFile(config: config);