+++ pkgs/native_toolchain_c/test/cbuilder/cbuilder_cross_android_test.dart
@@ -135,6 +135,7 @@ Future<Uri> buildLib(
     linkModePreference: linkMode == DynamicLoadingBundled()
         ? LinkModePreference.dynamic
         : LinkModePreference.static,
+    hasLinkPhase: false,
   );
   final buildOutput = BuildOutput();