+++ pkgs/native_toolchain_c/test/cbuilder/cbuilder_cross_ios_test.dart
@@ -73,6 +73,7 @@ void main() {
                     ? LinkModePreference.dynamic
                     : LinkModePreference.static,
                 targetIOSSdk: targetIOSSdk,
+                hasLinkPhase: false,
               );
               final buildOutput = BuildOutput();
 
@@ -209,6 +210,7 @@ Future<Uri> buildLib(
     linkModePreference: linkMode == DynamicLoadingBundled()
         ? LinkModePreference.dynamic
         : LinkModePreference.static,
+    hasLinkPhase: false,
   );
   final buildOutput = BuildOutput();