+++ pkgs/native_toolchain_c/test/cbuilder/cbuilder_cross_linux_host_test.dart
@@ -54,6 +54,7 @@ void main() {
           linkModePreference: linkMode == DynamicLoadingBundled()
               ? LinkModePreference.dynamic
               : LinkModePreference.static,
+          hasLinkPhase: false,
         );
         final buildOutput = BuildOutput();